+++ pkgs/test/test/runner/pub_serve_test.dart
@@ -9,10 +9,8 @@
 import 'dart:io';
 
 import 'package:path/path.dart' as p;
-import 'package:pub_semver/pub_semver.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
 
-import 'package:test_core/src/util/io.dart';
 import 'package:test/test.dart';
 
 import '../io.dart';
@@ -340,23 +338,14 @@ void main() {
 }
 
 /// The list of supported compilers for the current [Platform.version].
-final Iterable<String> _compilers = () {
-  var compilers = ['dart2js'];
-  if (_sdkSupportsDartDevc) compilers.add('dartdevc');
-  return compilers;
-}();
-
-/// Whether or not the dartdevc compiler is supported on the current
-/// [Platform.version].
-final bool _sdkSupportsDartDevc = sdkVersion >= Version(1, 24, 0);
+final Iterable<String> _compilers = ['dart2js', 'dartdevc'];
 
 /// Runs the test described by [testFn] once for each supported compiler on the
 /// current [Platform.version], passing the relevant compiler args for pub serve
 /// as the first argument.
 void testWithCompiler(String name, testFn(List<String> compilerArgs), {tags}) {
   for (var compiler in _compilers) {
-    var compilerArgs =
-        _sdkSupportsDartDevc ? ['--web-compiler', compiler] : <String>[];
+    var compilerArgs = ['--web-compiler', compiler];
     test("$name with $compiler", () => testFn(compilerArgs), tags: tags);
   }
 }