+++ pkgs/test_core/lib/src/runner/vm/environment.dart
@@ -5,7 +5,7 @@
 import 'dart:async';
 
 import 'package:async/async.dart';
-import 'package:vm_service_lib/vm_service_lib.dart';
+import 'package:vm_service/vm_service.dart';
 import 'package:test_core/src/runner/environment.dart'; // ignore: implementation_imports
 
 /// The environment in which VM tests are loaded.